refactor: start breaking up the update stuff
